rename internal packages per review feedback
- internal/context -> internal/contextkey (no longer shadows stdlib)
- internal/constant -> internal/constants (simple, clear package name)
- remove unused DefaultHeaderUsername constant
