+++ env.go
@@ -4,9 +4,9 @@ import (
 	"cmp"
 	"os"
 
-	"github.com/konflux-ci/namespace-lister/internal/constant"
+	"github.com/konflux-ci/namespace-lister/internal/constants"
 )
 
 func getAddress() string {
-	return cmp.Or(os.Getenv(constant.EnvAddress), constant.DefaultAddr)
+	return cmp.Or(os.Getenv(constants.EnvAddress), constants.DefaultAddr)
 }
