+++ internal/log/log.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/konflux-ci/namespace-lister/internal/constant"
-	nscontext "github.com/konflux-ci/namespace-lister/internal/context"
+	"github.com/konflux-ci/namespace-lister/internal/constants"
+	"github.com/konflux-ci/namespace-lister/internal/contextkey"
 )
 
 // BuildLogger constructs a new instance of the logger
@@ -23,7 +23,7 @@ func BuildLogger() *slog.Logger {
 
 // GetLogLevel fetches the log level from the appropriate environment variable
 func GetLogLevel() slog.Level {
-	env := os.Getenv(constant.EnvLogLevel)
+	env := os.Getenv(constants.EnvLogLevel)
 	level, err := strconv.Atoi(env)
 	if err != nil {
 		return slog.LevelError
@@ -33,13 +33,13 @@ func GetLogLevel() slog.Level {
 
 // SetLoggerIntoContext sets the provided logger into the context
 func SetLoggerIntoContext(ctx context.Context, logger *slog.Logger) context.Context {
-	return context.WithValue(ctx, nscontext.ContextKeyLogger, logger)
+	return context.WithValue(ctx, contextkey.ContextKeyLogger, logger)
 }
 
 // GetLoggerFromContext retrieves the logger from the context.
 // If no logger is present in the context, it will return a DiscardAll logger.
 func GetLoggerFromContext(ctx context.Context) *slog.Logger {
-	if l, ok := ctx.Value(nscontext.ContextKeyLogger).(*slog.Logger); ok && l != nil {
+	if l, ok := ctx.Value(contextkey.ContextKeyLogger).(*slog.Logger); ok && l != nil {
 		return l
 	}
